refactor(e2e-suite): split metadata tests to two subfolders
