+++ suite/e2e/tests/metadata/legacy/dropbox-api-errors.test.ts
@@ -1,7 +1,7 @@
-import { AccountLabelId } from '../../support/enums/accountLabelId';
-import { expect, test } from '../../support/fixtures';
-import { MetadataProvider } from '../../support/mocks/metadataMock';
-import { createTestAnnotation } from '../../support/reporters/annotations';
+import { AccountLabelId } from '../../../support/enums/accountLabelId';
+import { expect, test } from '../../../support/fixtures';
+import { MetadataProvider } from '../../../support/mocks/metadataMock';
+import { createTestAnnotation } from '../../../support/reporters/annotations';
 
 test.describe('Dropbox API errors', { tag: ['@webOnly', '@T3W1', '@T3T1'] }, () => {
     test.use({
