+++ suite/e2e/tests/metadata/legacy/switching-providers.test.ts
@@ -1,7 +1,7 @@
-import { AccountLabelId } from '../../support/enums/accountLabelId';
-import { expect, test } from '../../support/fixtures';
-import { MetadataProvider } from '../../support/mocks/metadataMock';
-import { createTestAnnotation } from '../../support/reporters/annotations';
+import { AccountLabelId } from '../../../support/enums/accountLabelId';
+import { expect, test } from '../../../support/fixtures';
+import { MetadataProvider } from '../../../support/mocks/metadataMock';
+import { createTestAnnotation } from '../../../support/reporters/annotations';
 
 test.describe(
     'Metadata - switching between cloud providers',
